Separate Timeout for Screen and Security Lock (UI)
This allows the device to engage the slide-to-unlock lock screen after a delay followed by automatic transition to the security lock (ie. Pattern/PIN/Password) screen after a further delay.

Control is provided for distinct slide lock screen timeout and screen turned off delays. If the feature is disabled or the slide lock delay is equal to or greater than the corresponding security lock delay (the default case), the slide lock screen will never be encountered and only the security lock screen will be displayed.

This patchset attempts to maintain conventional behavior of the security lock screens. For example, in the case of Pattern lock screen security the slide lock screen will be replaced by the Pattern lock screen after the relevant delay for the security lock, so that only the pattern lock screen will be displayed once it is engaged. All other types of security lock screens are displayed after the slide lock screen has been disengaged. The same behavior holds for boot: the pattern lock screen will be displayed immediately, and all others will be presented as slide -> security.

UI for the feature is located in CyanogenMod settings -> Lockscreen -> Delay and Timeout. The feature is disabled by default. Also, if the device has security lock type set to "None", a note is displayed to explain the configuration semantics.

Change-Id: Ica797b980331739f28b8d1e983829e2484707b1e